+++ contracts/lib/protobuf/SupplementalLineItem.sol
@@ -113,7 +113,7 @@ function decode_State(int64 x) internal pure returns (State) {
    * @return The decoded struct
    */
   function decode(bytes memory bs) internal pure returns (Data memory) {
-    (Data memory x,) = _decode(32, bs, bs.length);
+    (Data memory x, ) = _decode(32, bs, bs.length);
     return x;
   }
 
@@ -123,7 +123,7 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param bs The bytes array to be decoded
    */
   function decode(Data storage self, bytes memory bs) internal {
-    (Data memory x,) = _decode(32, bs, bs.length);
+    (Data memory x, ) = _decode(32, bs, bs.length);
     store(x, self);
   }
   // inner decoder
@@ -136,64 +136,67 @@ function decode_State(int64 x) internal pure returns (State) {
    * @return The decoded struct
    * @return The number of bytes decoded
    */
-  function _decode(uint p, bytes memory bs, uint sz)
-      internal pure returns (Data memory, uint) {
+  function _decode(uint256 p, bytes memory bs, uint256 sz)
+    internal 
+    pure 
+    returns (Data memory, uint) 
+  {
     Data memory r;
     uint[10] memory counters;
-    uint fieldId;
+    uint256 fieldId;
     ProtoBufRuntime.WireType wireType;
-    uint bytesRead;
-    uint offset = p;
-    uint pointer = p;
-    while(pointer < offset+sz) {
+    uint256 bytesRead;
+    uint256 offset = p;
+    uint256 pointer = p;
+    while (pointer < offset + sz) {
       (fieldId, wireType, bytesRead) = ProtoBufRuntime._decode_key(pointer, bs);
       pointer += bytesRead;
-      if(fieldId == 1) {
+      if (fieldId == 1) {
         pointer += _read_id(pointer, bs, r, counters);
       }
-      else if(fieldId == 2) {
+      else if (fieldId == 2) {
         pointer += _read_lineItemType(pointer, bs, r, counters);
       }
-      else if(fieldId == 3) {
+      else if (fieldId == 3) {
         pointer += _read_state(pointer, bs, r, counters);
       }
-      else if(fieldId == 4) {
+      else if (fieldId == 4) {
         pointer += _read_obligatorAddress(pointer, bs, r, counters);
       }
-      else if(fieldId == 5) {
+      else if (fieldId == 5) {
         pointer += _read_claimorAddress(pointer, bs, r, counters);
       }
-      else if(fieldId == 6) {
+      else if (fieldId == 6) {
         pointer += _read_tokenAddress(pointer, bs, r, counters);
       }
-      else if(fieldId == 7) {
+      else if (fieldId == 7) {
         pointer += _read_amount(pointer, bs, r, counters);
       }
-      else if(fieldId == 8) {
+      else if (fieldId == 8) {
         pointer += _read_dueTimestamp(pointer, bs, r, counters);
       }
-      else if(fieldId == 9) {
+      else if (fieldId == 9) {
         pointer += _read_reinitiatedTo(pointer, bs, r, counters);
       }
       
       else {
         if (wireType == ProtoBufRuntime.WireType.Fixed64) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_fixed64(pointer, bs);
           pointer += size;
         }
         if (wireType == ProtoBufRuntime.WireType.Fixed32) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_fixed32(pointer, bs);
           pointer += size;
         }
         if (wireType == ProtoBufRuntime.WireType.Varint) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_varint(pointer, bs);
           pointer += size;
         }
         if (wireType == ProtoBufRuntime.WireType.LengthDelim) {
-          uint size;
+          uint256 size;
           (, size) = ProtoBufRuntime._decode_lendelim(pointer, bs);
           pointer += size;
         }
@@ -213,16 +216,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_id(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_id(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (uint8 x, uint sz) = ProtoBufRuntime._decode_sol_uint8(p, bs);
-    if(isNil(r)) {
+    (uint8 x, uint256 sz) = ProtoBufRuntime._decode_sol_uint8(p, bs);
+    if (isNil(r)) {
       counters[1] += 1;
     } else {
       r.id = x;
-      if(counters[1] > 0) counters[1] -= 1;
+      if (counters[1] > 0) counters[1] -= 1;
     }
     return sz;
   }
@@ -235,13 +243,18 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_lineItemType(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_lineItemType(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (int64 tmp, uint sz) = ProtoBufRuntime._decode_enum(p, bs);
+    (int64 tmp, uint256 sz) = ProtoBufRuntime._decode_enum(p, bs);
     SupplementalLineItem.Type x = decode_Type(tmp);
-    if(isNil(r)) {
+    if (isNil(r)) {
       counters[2] += 1;
     } else {
       r.lineItemType = x;
@@ -258,13 +271,18 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_state(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_state(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (int64 tmp, uint sz) = ProtoBufRuntime._decode_enum(p, bs);
+    (int64 tmp, uint256 sz) = ProtoBufRuntime._decode_enum(p, bs);
     SupplementalLineItem.State x = decode_State(tmp);
-    if(isNil(r)) {
+    if (isNil(r)) {
       counters[3] += 1;
     } else {
       r.state = x;
@@ -281,16 +299,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_obligatorAddress(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_obligatorAddress(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (address x, uint sz) = ProtoBufRuntime._decode_sol_address(p, bs);
-    if(isNil(r)) {
+    (address x, uint256 sz) = ProtoBufRuntime._decode_sol_address(p, bs);
+    if (isNil(r)) {
       counters[4] += 1;
     } else {
       r.obligatorAddress = x;
-      if(counters[4] > 0) counters[4] -= 1;
+      if (counters[4] > 0) counters[4] -= 1;
     }
     return sz;
   }
@@ -303,16 +326,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_claimorAddress(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_claimorAddress(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (address x, uint sz) = ProtoBufRuntime._decode_sol_address(p, bs);
-    if(isNil(r)) {
+    (address x, uint256 sz) = ProtoBufRuntime._decode_sol_address(p, bs);
+    if (isNil(r)) {
       counters[5] += 1;
     } else {
       r.claimorAddress = x;
-      if(counters[5] > 0) counters[5] -= 1;
+      if (counters[5] > 0) counters[5] -= 1;
     }
     return sz;
   }
@@ -325,16 +353,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_tokenAddress(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_tokenAddress(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (address x, uint sz) = ProtoBufRuntime._decode_sol_address(p, bs);
-    if(isNil(r)) {
+    (address x, uint256 sz) = ProtoBufRuntime._decode_sol_address(p, bs);
+    if (isNil(r)) {
       counters[6] += 1;
     } else {
       r.tokenAddress = x;
-      if(counters[6] > 0) counters[6] -= 1;
+      if (counters[6] > 0) counters[6] -= 1;
     }
     return sz;
   }
@@ -347,16 +380,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_amount(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_amount(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (uint256 x, uint sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
-    if(isNil(r)) {
+    (uint256 x, uint256 sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
+    if (isNil(r)) {
       counters[7] += 1;
     } else {
       r.amount = x;
-      if(counters[7] > 0) counters[7] -= 1;
+      if (counters[7] > 0) counters[7] -= 1;
     }
     return sz;
   }
@@ -369,16 +407,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_dueTimestamp(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_dueTimestamp(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (uint256 x, uint sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
-    if(isNil(r)) {
+    (uint256 x, uint256 sz) = ProtoBufRuntime._decode_sol_uint256(p, bs);
+    if (isNil(r)) {
       counters[8] += 1;
     } else {
       r.dueTimestamp = x;
-      if(counters[8] > 0) counters[8] -= 1;
+      if (counters[8] > 0) counters[8] -= 1;
     }
     return sz;
   }
@@ -391,16 +434,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param counters The counters for repeated fields
    * @return The number of bytes decoded
    */
-  function _read_reinitiatedTo(uint p, bytes memory bs, Data memory r, uint[10] memory counters) internal pure returns (uint) {
+  function _read_reinitiatedTo(
+    uint256 p, 
+    bytes memory bs, 
+    Data memory r, 
+    uint[10] memory counters
+  ) internal pure returns (uint) {
     /**
      * if `r` is NULL, then only counting the number of fields.
      */
-    (uint8 x, uint sz) = ProtoBufRuntime._decode_sol_uint8(p, bs);
-    if(isNil(r)) {
+    (uint8 x, uint256 sz) = ProtoBufRuntime._decode_sol_uint8(p, bs);
+    if (isNil(r)) {
       counters[9] += 1;
     } else {
       r.reinitiatedTo = x;
-      if(counters[9] > 0) counters[9] -= 1;
+      if (counters[9] > 0) counters[9] -= 1;
     }
     return sz;
   }
@@ -415,7 +463,7 @@ function decode_State(int64 x) internal pure returns (State) {
    */
   function encode(Data memory r) internal pure returns (bytes memory) {
     bytes memory bs = new bytes(_estimate(r));
-    uint sz = _encode(r, 32, bs);
+    uint256 sz = _encode(r, 32, bs);
     assembly {
       mstore(bs, sz)
     }
@@ -430,30 +478,78 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param bs The bytes array to be decoded
    * @return The number of bytes encoded
    */
-  function _encode(Data memory r, uint p, bytes memory bs)
-      internal pure returns (uint) {
-    uint offset = p;
-    uint pointer = p;
+  function _encode(Data memory r, uint256 p, bytes memory bs)
+    internal 
+    pure 
+    returns (uint) 
+  {
+    uint256 offset = p;
+    uint256 pointer = p;
     
-    pointer += ProtoBufRuntime._encode_key(1, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      1, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_uint8(r.id, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(2, ProtoBufRuntime.WireType.Varint, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      2, 
+      ProtoBufRuntime.WireType.Varint, 
+      pointer, 
+      bs
+    );
     int64 _enum_lineItemType = encode_Type(r.lineItemType);
     pointer += ProtoBufRuntime._encode_enum(_enum_lineItemType, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(3, ProtoBufRuntime.WireType.Varint, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      3, 
+      ProtoBufRuntime.WireType.Varint, 
+      pointer, 
+      bs
+    );
     int64 _enum_state = encode_State(r.state);
     pointer += ProtoBufRuntime._encode_enum(_enum_state, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(4, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      4, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_address(r.obligatorAddress, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(5, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      5, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_address(r.claimorAddress, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(6, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      6, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_address(r.tokenAddress, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(7, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      7, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_uint256(r.amount, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(8, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      8, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_uint256(r.dueTimestamp, pointer, bs);
-    pointer += ProtoBufRuntime._encode_key(9, ProtoBufRuntime.WireType.LengthDelim, pointer, bs);
+    pointer += ProtoBufRuntime._encode_key(
+      9, 
+      ProtoBufRuntime.WireType.LengthDelim, 
+      pointer, 
+      bs
+    );
     pointer += ProtoBufRuntime._encode_sol_uint8(r.reinitiatedTo, pointer, bs);
     return pointer - offset;
   }
@@ -466,18 +562,21 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param bs The bytes array to be decoded
    * @return The number of bytes encoded
    */
-  function _encode_nested(Data memory r, uint p, bytes memory bs)
-      internal pure returns (uint) {
+  function _encode_nested(Data memory r, uint256 p, bytes memory bs)
+    internal 
+    pure 
+    returns (uint) 
+  {
     /**
      * First encoded `r` into a temporary array, and encode the actual size used.
      * Then copy the temporary array into `bs`.
      */
-    uint offset = p;
-    uint pointer = p;
+    uint256 offset = p;
+    uint256 pointer = p;
     bytes memory tmp = new bytes(_estimate(r));
-    uint tmpAddr = ProtoBufRuntime.getMemoryAddress(tmp);
-    uint bsAddr = ProtoBufRuntime.getMemoryAddress(bs);
-    uint size = _encode(r, 32, tmp);
+    uint256 tmpAddr = ProtoBufRuntime.getMemoryAddress(tmp);
+    uint256 bsAddr = ProtoBufRuntime.getMemoryAddress(bs);
+    uint256 size = _encode(r, 32, tmp);
     pointer += ProtoBufRuntime._encode_varint(size, pointer, bs);
     ProtoBufRuntime.copyBytes(tmpAddr + 32, bsAddr + pointer, size);
     pointer += size;
@@ -491,8 +590,10 @@ function decode_State(int64 x) internal pure returns (State) {
    * @param r The struct to be encoded
    * @return The number of bytes encoded in estimation
    */
-  function _estimate(Data memory r) internal pure returns (uint) {
-    uint e;
+  function _estimate(
+    Data memory r
+  ) internal pure returns (uint) {
+    uint256 e;
     e += 1 + 4;
     e += 1 + ProtoBufRuntime._sz_enum(encode_Type(r.lineItemType));
     e += 1 + ProtoBufRuntime._sz_enum(encode_State(r.state));
